move test from test_storage to test_util
